% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve function}
\usage{
myncurve(mu, sigma, le, re, lee, ree)
}
\arguments{
\item{mu}{Mean of the normal function}

\item{sigma}{Standard deviation of the normal function}

\item{le}{Left endpoint}

\item{re}{Right endpoint}

\item{lee}{Does the line go on the left endpoint?}

\item{ree}{Does the line go on the right endpoint?}
}
\description{
prints out a graph, region of the area we are looking for, and area itself
}
